freeweight machines in Qatar
The Green Sports principal trading entity, established in 2016 has developed a solid reputation in Qatar for its marketing abilities & long term commitment to a varied clientele.
We are fitness & Spa equipment supplier with a futuristic outlook and commitment to the customer, which is of prime importance to us in building enduring relationships for the future. Green Sports department specializes in supply, installation & maintenance of high quality fitness and spa equipment for commercial & Home gyms, health clubs, hotels, sports clubs, military facilities etc. The company has a long history of success and has been involved in many of the most prestigious fitness facilities developments in the region.